// src/a2aclient/types.ts
export type A2AError =
  | JSONParseError
  | InvalidRequestError
  | MethodNotFoundError
  | InvalidParamsError
  | InternalError
  | TaskNotFoundError
  | TaskNotCancelableError
  | PushNotificationNotSupportedError
  | UnsupportedOperationError
  | ContentTypeNotSupportedError
  | InvalidAgentResponseError;
// ...
export type Part = TextPart | FilePart | DataPart;
/**
 * Mirrors the OpenAPI Security Scheme Object
 * (https://swagger.io/specification/#security-scheme-object)
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SecurityScheme".
 */
export type SecurityScheme =
  | APIKeySecurityScheme
  | HTTPAuthSecurityScheme
  | OAuth2SecurityScheme
  | OpenIdConnectSecurityScheme;
/**
 * JSON-RPC response for the 'tasks/cancel' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "CancelTaskResponse".
 */
export type CancelTaskResponse = JSONRPCErrorResponse | CancelTaskSuccessResponse;
// ...
export type GetTaskPushNotificationConfigResponse = JSONRPCErrorResponse | GetTaskPushNotificationConfigSuccessResponse;
/**
 * JSON-RPC response for the 'tasks/get' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "GetTaskResponse".
 */
export type GetTaskResponse = JSONRPCErrorResponse | GetTaskSuccessResponse;
/**
 * Represents a JSON-RPC 2.0 Response object.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "JSONRPCResponse".
 */
export type JSONRPCResponse =
  | JSONRPCErrorResponse
  | SendMessageSuccessResponse
  | SendStreamingMessageSuccessResponse
  | GetTaskSuccessResponse
  | CancelTaskSuccessResponse
  | SetTaskPushNotificationConfigSuccessResponse
  | GetTaskPushNotificationConfigSuccessResponse;
/**
 * JSON-RPC response model for the 'message/send' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SendMessageResponse".
 */
export type SendMessageResponse = JSONRPCErrorResponse | SendMessageSuccessResponse;
/**
 * JSON-RPC response model for the 'message/stream' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SendStreamingMessageResponse".
 */
export type SendStreamingMessageResponse = JSONRPCErrorResponse | SendStreamingMessageSuccessResponse;
/**
 * JSON-RPC response for the 'tasks/pushNotificationConfig/set' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SetTaskPushNotificationConfigResponse".
 */
export type SetTaskPushNotificationConfigResponse = JSONRPCErrorResponse | SetTaskPushNotificationConfigSuccessResponse;

export interface MySchema {
  [k: string]: unknown;
}
/**
 * JSON-RPC error indicating invalid JSON was received by the server.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "JSONParseError".
 */
export interface JSONParseError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32700;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC error indicating the JSON sent is not a valid Request object.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "InvalidRequestError".
 */
export interface InvalidRequestError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: -32600;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
// ...
export interface MessageSendParams {
  configuration?: MessageSendConfiguration;
  message: Message;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
}
// ...
export interface Message {
  /**
   * The context the message is associated with
   */
  contextId?: string;
  /**
   * The URIs of extensions that are present or contributed to this Message.
   */
  extensions?: string[];
  /**
   * Event type
   */
  kind: "message";
  /**
   * Identifier created by the message creator
   */
  messageId: string;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Message content
   */
  parts: Part[];
  /**
   * List of tasks referenced as context by this message.
   */
  referenceTaskIds?: string[];
  /**
   * Message sender's role
   */
  role: "agent" | "user";
  /**
   * Identifier of task the message is related to
   */
  taskId?: string;
}
// ...
export interface TaskQueryParams {
  /**
   * Number of recent messages to be retrieved.
   */
  historyLength?: number;
  /**
   * Task id.
   */
  id: string;
  metadata?: {
    [k: string]: unknown;
  };
}
// ...
export interface TaskIdParams {
  /**
   * Task id.
   */
  id: string;
  metadata?: {
    [k: string]: unknown;
  };
}
// ...
export interface TaskPushNotificationConfig {
  pushNotificationConfig: PushNotificationConfig1;
  /**
   * Task id.
   */
  taskId: string;
}
// ...
export interface AgentCard {
  capabilities: AgentCapabilities1;
  /**
   * The set of interaction modes that the agent supports across all skills. This can be overridden per-skill.
   * Supported media types for input.
   */
  defaultInputModes: string[];
  /**
   * Supported media types for output.
   */
  defaultOutputModes: string[];
  /**
   * A human-readable description of the agent. Used to assist users and
   * other agents in understanding what the agent can do.
   */
  description: string;
  /**
   * A URL to documentation for the agent.
   */
  documentationUrl?: string;
  /**
   * A URL to an icon for the agent.
   */
  iconUrl?: string;
  /**
   * Human readable name of the agent.
   */
  name: string;
  provider?: AgentProvider;
  /**
   * Security requirements for contacting the agent.
   */
  security?: {
    [k: string]: string[];
  }[];
  /**
   * Security scheme details used for authenticating with this agent.
   */
  securitySchemes?: {
    [k: string]: SecurityScheme;
  };
  /**
   * Skills are a unit of capability that an agent can perform.
   */
  skills: AgentSkill[];
  /**
   * true if the agent supports providing an extended agent card when the user is authenticated.
   * Defaults to false if not specified.
   */
  supportsAuthenticatedExtendedCard?: boolean;
  /**
   * A URL to the address the agent is hosted at.
   */
  url: string;
  /**
   * The version of the agent - format is up to the provider.
   */
  version: string;
}
/**
 * Optional capabilities supported by the agent.
 */
export interface AgentCapabilities1 {
  /**
   * extensions supported by this agent.
   */
  extensions?: AgentExtension[];
  /**
   * true if the agent can notify updates to client.
   */
  pushNotifications?: boolean;
  /**
   * true if the agent exposes status change history for tasks.
   */
  stateTransitionHistory?: boolean;
  /**
   * true if the agent supports SSE.
   */
  streaming?: boolean;
}
/**
 * The service provider of the agent
 */
export interface AgentProvider {
  /**
   * Agent provider's organization name.
   */
  organization: string;
  /**
   * Agent provider's URL.
   */
  url: string;
}
// ...
export interface AgentSkill {
  /**
   * Description of the skill - will be used by the client or a human
   * as a hint to understand what the skill does.
   */
  description: string;
  /**
   * The set of example scenarios that the skill can perform.
   * Will be used by the client as a hint to understand how the skill can be used.
   */
  examples?: string[];
  /**
   * Unique identifier for the agent's skill.
   */
  id: string;
  /**
   * The set of interaction modes that the skill supports
   * (if different than the default).
   * Supported media types for input.
   */
  inputModes?: string[];
  /**
   * Human readable name of the skill.
   */
  name: string;
  /**
   * Supported media types for output.
   */
  outputModes?: string[];
  /**
   * Set of tagwords describing classes of capabilities for this specific skill.
   */
  tags: string[];
}
// ...
export interface JSONRPCErrorResponse {
  error:
    | JSONRPCError
    | JSONParseError
    | InvalidRequestError
    | MethodNotFoundError
    | InvalidParamsError
    | InternalError
    | TaskNotFoundError
    | TaskNotCancelableError
    | PushNotificationNotSupportedError
    | UnsupportedOperationError
    | ContentTypeNotSupportedError
    | InvalidAgentResponseError;
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
}
/**
 * Represents a JSON-RPC 2.0 Error object.
 * This is typically included in a JSONRPCErrorResponse when an error occurs.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "JSONRPCError".
 */
export interface JSONRPCError {
  /**
   * A Number that indicates the error type that occurred.
   */
  code: number;
  /**
   * A Primitive or Structured value that contains additional information about the error.
   * This may be omitted.
   */
  data?: {
    [k: string]: unknown;
  };
  /**
   * A String providing a short description of the error.
   */
  message: string;
}
/**
 * JSON-RPC success response model for the 'tasks/cancel' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "CancelTaskSuccessResponse".
 */
export interface CancelTaskSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  result: Task;
}
/**
 * The result object on success.
 */
export interface Task {
  /**
   * Collection of artifacts created by the agent.
   */
  artifacts?: Artifact[];
  /**
   * Server-generated id for contextual alignment across interactions
   */
  contextId: string;
  history?: Message1[];
  /**
   * Unique identifier for the task
   */
  id: string;
  /**
   * Event type
   */
  kind: "task";
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  status: TaskStatus;
}
// ...
export interface TaskStatus {
  message?: Message2;
  state: TaskState;
  /**
   * ISO 8601 datetime string when the status was recorded.
   */
  timestamp?: string;
}
// ...
export interface GetTaskPushNotificationConfigSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  result: TaskPushNotificationConfig1;
}
/**
 * The result object on success.
 */
export interface TaskPushNotificationConfig1 {
  pushNotificationConfig: PushNotificationConfig1;
  /**
   * Task id.
   */
  taskId: string;
}
/**
 * JSON-RPC success response for the 'tasks/get' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "GetTaskSuccessResponse".
 */
export interface GetTaskSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  result: Task1;
}
// ...
export interface JSONRPCRequest {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id?: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  /**
   * A String containing the name of the method to be invoked.
   */
  method: string;
  /**
   * A Structured value that holds the parameter values to be used during the invocation of the method.
   */
  params?: {
    [k: string]: unknown;
  };
}
/**
 * JSON-RPC success response model for the 'message/send' method.
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "SendMessageSuccessResponse".
 */
export interface SendMessageSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  /**
   * The result object on success
   */
  result: Task2 | Message1;
}
// ...
export interface SendStreamingMessageSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  /**
   * The result object on success
   */
  result: Task2 | Message1 | TaskStatusUpdateEvent | TaskArtifactUpdateEvent;
}
/**
 * Sent by server during sendStream or subscribe requests
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "TaskStatusUpdateEvent".
 */
export interface TaskStatusUpdateEvent {
  /**
   * The context the task is associated with
   */
  contextId: string;
  /**
   * Indicates the end of the event stream
   */
  final: boolean;
  /**
   * Event type
   */
  kind: "status-update";
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  status: TaskStatus1;
  /**
   * Task id
   */
  taskId: string;
}
/**
 * Current status of the task
 */
export interface TaskStatus1 {
  message?: Message2;
  state: TaskState;
  /**
   * ISO 8601 datetime string when the status was recorded.
   */
  timestamp?: string;
}
/**
 * Sent by server during sendStream or subscribe requests
 *
 * This interface was referenced by `MySchema`'s JSON-Schema
 * via the `definition` "TaskArtifactUpdateEvent".
 */
export interface TaskArtifactUpdateEvent {
  /**
   * Indicates if this artifact appends to a previous one
   */
  append?: boolean;
  artifact: Artifact1;
  /**
   * The context the task is associated with
   */
  contextId: string;
  /**
   * Event type
   */
  kind: "artifact-update";
  /**
   * Indicates if this is the last chunk of the artifact
   */
  lastChunk?: boolean;
  /**
   * Extension metadata.
   */
  metadata?: {
    [k: string]: unknown;
  };
  /**
   * Task id
   */
  taskId: string;
}
// ...
export interface SetTaskPushNotificationConfigSuccessResponse {
  /**
   * An identifier established by the Client that MUST contain a String, Number.
   * Numbers SHOULD NOT contain fractional parts.
   */
  id: string | number | null;
  /**
   * Specifies the version of the JSON-RPC protocol. MUST be exactly "2.0".
   */
  jsonrpc: "2.0";
  result: TaskPushNotificationConfig2;
}